login-required decorator, fixed token cache, context adapters re-write
