+++ templates/auth/status.html
@@ -6,10 +6,10 @@
 
 {% block text %}
     {% if ms_id_user_principal.authenticated %}
-        Click here to get your <a class="btn btn-success p-1" href="{{ url_for('auth.token_details') }}">ID Token Details</a>
+        Click here to get your <a class="btn btn-success p-1" href="{{ url_for('token_details') }}">ID Token Details</a>
         or <a class="btn btn-success p-1" href="{{ url_for('auth.edit_profile') }}">Edit Profile</a>
     {% else %}
         Use the button on the top right to sign in. <br/>
-        Attempts to get your <a href="{{ url_for('auth.token_details') }}">ID Token Details</a> should result in a 401 error.
+        Attempts to get your <a href="{{ url_for('token_details') }}">ID Token Details</a> should result in a 401 error.
     {% endif %}
 {% endblock %}
